# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elicta Service",
    "version": "0.1.0"
  },
  "paths": {
    "/api/meetings/{meeting_id}/consent-gate": {
      "get": {
        "tags": [
          "consent"
        ],
        "summary": "Get Consent Gate",
        "operationId": "get_consent_gate_api_meetings__meeting_id__consent_gate_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          },
          {
            "name": "engagement_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentGate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/consent-confirmation": {
      "post": {
        "tags": [
          "consent"
        ],
        "summary": "Confirm Consent",
        "operationId": "confirm_consent_api_meetings__meeting_id__consent_confirmation_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentConfirmationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagements": {
      "post": {
        "tags": [
          "engagements"
        ],
        "summary": "Create Engagement Endpoint",
        "operationId": "create_engagement_endpoint_api_engagements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Create Meeting Endpoint",
        "operationId": "create_meeting_endpoint_api_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{meeting_id}/bank": {
      "get": {
        "tags": [
          "compiler-bank"
        ],
        "summary": "Get Meeting Bank",
        "operationId": "get_meeting_bank_api_meetings__meeting_id__bank_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingQuestionBank"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagements/{engagement_id}/bank": {
      "get": {
        "tags": [
          "compiler-bank"
        ],
        "summary": "Get Engagement Bank",
        "operationId": "get_engagement_bank_api_engagements__engagement_id__bank_get",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementQuestionBank"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}/bank/compile": {
      "post": {
        "tags": [
          "compiler-bank"
        ],
        "summary": "Compile Engagement Bank",
        "operationId": "compile_engagement_bank_api_engagements__engagement_id__bank_compile_post",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankCompileTrigger"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bank/candidates/{candidate_id}": {
      "delete": {
        "tags": [
          "compiler-bank"
        ],
        "summary": "Delete Bank Candidate",
        "operationId": "delete_bank_candidate_api_bank_candidates__candidate_id__delete",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "compiler-bank"
        ],
        "summary": "Patch Bank Candidate",
        "operationId": "patch_bank_candidate_api_bank_candidates__candidate_id__patch",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidatePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__modules__compiler__api__models__BankCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BankCompileTrigger": {
        "properties": {
          "job_id": {
            "type": "string",
            "minLength": 1,
            "title": "Job Id"
          },
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "triggered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Triggered At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "engagement_id",
          "triggered_at"
        ],
        "title": "BankCompileTrigger",
        "description": "Handle for one engagement's bank compile job, returned immediately on trigger (PRD FR-4.8).\n\nReturned when `POST /api/engagements/{id}/bank/compile` is accepted,\nbefore the actual compiling work (extraction, tagging, embedding -- all\nout of this feature's footprint) has produced any candidates yet.\n`job_id` is what a caller polls or correlates against later, mirroring\n`RecordPathTranscriptionJob`'s (`asr-record/models.py`) job-handle\nconvention for an accepted, not-yet-finished run."
      },
...
      "CandidatePatchRequest": {
        "properties": {
          "phrasing": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Phrasing"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "pruned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pruned"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CandidatePatchRequest",
        "description": "A partial update to one bank candidate: edit its phrasing, reorder its priority, or prune it (PRD FR-4.8).\n\nAll fields are optional so a caller can change just one aspect at a\ntime, but at least one must be supplied, since a PATCH with nothing to\nchange has no meaningful effect to report as a 200."
      },
...
      "ConsentConfirmationRequest": {
        "properties": {
          "confirmed_by": {
            "type": "string",
            "minLength": 1,
            "title": "Confirmed By"
          }
        },
        "type": "object",
        "required": [
          "confirmed_by"
        ],
        "title": "ConsentConfirmationRequest",
        "description": "Payload for an operator confirming the consent prompt (PRD feature 246)."
      },
      "ConsentGate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConsentGateStatus"
          },
          "prompt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentPrompt"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ConsentGate",
        "description": "Result of evaluating whether a meeting may begin capture.\n\n``capture_may_begin`` is the single field a capture-start flow needs to\ncheck. ``prompt`` is populated only when the operator must confirm\nconsent before capture is allowed to proceed."
      },
      "ConsentGateStatus": {
        "type": "string",
        "enum": [
          "not_required",
          "awaiting_confirmation",
          "confirmed"
        ],
        "title": "ConsentGateStatus",
        "description": "Where a given meeting sits relative to the consent gate."
      },
      "ConsentPrompt": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "legal_basis": {
            "type": "string",
            "title": "Legal Basis"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body",
          "legal_basis"
        ],
        "title": "ConsentPrompt",
        "description": "Operator-facing copy shown before capture begins (PRD L2)."
      },
      "ConsentRecord": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "confirmed_by": {
            "type": "string",
            "title": "Confirmed By"
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "confirmed_by",
          "confirmed_at"
        ],
        "title": "ConsentRecord",
        "description": "Durable proof that consent was confirmed for one meeting (PRD feature 246).\n\n``confirmed_by`` identifies the operator who confirmed the prompt, not\nthe meeting participants \u2014 this records accountability for having\ndisclosed and confirmed consent, which is what a later audit needs."
      },
...
      "EngagementContext": {
        "properties": {
          "client_organisation": {
            "type": "string",
            "title": "Client Organisation"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "commercial_context": {
            "type": "string",
            "title": "Commercial Context"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "scope_boundary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Boundary"
          },
          "target_requirements_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Requirements Template"
          }
        },
        "type": "object",
        "required": [
          "client_organisation",
          "sector",
          "commercial_context"
        ],
        "title": "EngagementContext",
        "description": "The engagement-level context a new meeting inherits automatically (PRD FR-3.1, FR-3.5, FR-3.7).\n\nMirrors the engagement's own context fields -- three captured at\nengagement creation (`client_organisation`, `sector`,\n`commercial_context`), three filled in later (`purpose`,\n`scope_boundary`, `target_requirements_template`) -- so a meeting-create\nresponse can echo them back for confirmation without this package\nimporting the engagement package's persistence-backed record type."
      },
      "EngagementCreateRequest": {
        "properties": {
          "client_organisation": {
            "type": "string",
            "minLength": 1,
            "title": "Client Organisation"
          },
          "sector": {
            "type": "string",
            "minLength": 1,
            "title": "Sector"
          },
          "commercial_context": {
            "type": "string",
            "minLength": 1,
            "title": "Commercial Context"
          }
        },
        "type": "object",
        "required": [
          "client_organisation",
          "sector",
          "commercial_context"
        ],
        "title": "EngagementCreateRequest",
        "description": "Client background captured once, at engagement creation (PRD FR-3.1)."
      },
      "EngagementCreateResponse": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          }
        },
        "type": "object",
        "required": [
          "engagement_id"
        ],
        "title": "EngagementCreateResponse"
      },
...
      "EngagementQuestionBank": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/QuestionBankSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "engagement_id",
          "sections",
          "generated_at"
        ],
        "title": "EngagementQuestionBank",
        "description": "An engagement's compiled question bank, rendered as a reviewable tree grouped by template section (PRD FR-4.8).\n\nGrouped into `sections` rather than returned as `MeetingQuestionBank`'s\nflat candidate list: this is the engagement's own compiled set (from\n`POST /api/engagements/{id}/bank/compile`, out of this feature's\nfootprint) as a reviewer inspects it once, before any individual\nmeeting recompiles from it -- so it renders as branches a reviewer can\nexpand and act on rather than one long list."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingCreateRequest": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "minLength": 1,
            "title": "Engagement Id"
          },
          "capture_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Capture Mode"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "engagement_id",
          "capture_mode"
        ],
        "title": "MeetingCreateRequest",
        "description": "What's specific to a new meeting -- its engagement, capture mode, and schedule (PRD FR-3.7, FR-3.8).\n\nDeliberately excludes every engagement-level context field: those are\ninherited automatically from `engagement_id` rather than re-entered\nhere. `capture_mode` is required since the `meetings` table declares it\n`NOT NULL` with no default; `scheduled_at` is optional since a meeting\ncan be created before a time is settled."
      },
      "MeetingCreateResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "capture_mode": {
            "type": "string",
            "title": "Capture Mode"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "engagement_context": {
            "$ref": "#/components/schemas/EngagementContext"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "engagement_id",
          "state",
          "capture_mode",
          "engagement_context"
        ],
        "title": "MeetingCreateResponse",
        "description": "A newly created meeting, with its inherited engagement context attached for confirmation (PRD FR-3.7)."
      },
...
      "MeetingQuestionBank": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/app__modules__compiler__bank__models__BankCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "candidates",
          "generated_at"
        ],
        "title": "MeetingQuestionBank",
        "description": "The fresh, per-meeting recompiled question bank (PRD FR-4.8).\n\nRecomputed by `recompile_meeting_bank` for every meeting rather than\nreused across an engagement's meetings \u2014 \"the bank\" a live meeting reads\nfrom is always this meeting's own recompile, not the engagement's\noriginal compile from `POST /api/engagements/{id}/bank/compile`."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__modules__compiler__api__models__BankCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "template_section": {
            "type": "string",
            "title": "Template Section"
          },
          "phrasing": {
            "type": "string",
            "title": "Phrasing"
          },
          "priority": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Priority"
          },
          "inherited_from_open_question": {
            "type": "boolean",
            "title": "Inherited From Open Question",
            "default": false
          },
          "pruned": {
            "type": "boolean",
            "title": "Pruned",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_section",
          "phrasing",
          "priority"
        ],
        "title": "BankCandidate",
        "description": "One candidate question in a meeting's recompiled bank (PRD FR-4.8, architecture section 3.6).\n\n`priority` orders candidates within the bank -- lower is ranked higher --\nmirroring the ascending `impact_rank` convention `OpenQuestion` already\nuses elsewhere in this codebase (`debrief/pipeline/models.py`).\n`inherited_from_open_question` is `True` only for a candidate\n`recompile_meeting_bank` derived directly from a prior meeting's open\nquestion rather than from the engagement's own compiled candidate set, so\na caller can render \"carried forward from last time\" distinctly from a\nfreshly compiled candidate. `pruned` is `True` once a caller has excluded\nthe candidate from the bank via `PATCH /api/bank/candidates/{id}` --\nwhoever supplies `get_base_candidates` (out of this feature's footprint)\nis responsible for leaving pruned candidates out of what it returns, the\nsame way it would leave out a hard-deleted one."
      },
...
    }
  }
}